# Conflicts:
#	VVS/VVS.Android/VVS.Android.csproj
#	VVS/VVS.Android/packages.config
#	VVS/VVS/Model/Replacement.cs
#	VVS/VVS/VVS.csproj
#	VVS/VVS/packages.config

+++ VVS/VVS/Model/Meter.cs
@@ -10,14 +10,14 @@ namespace VVS.Model
     public class Meter
     {
         [PrimaryKey]
-        public int SerialNumber { get; set; }
+        public long SerialNumber { get; set; }
         public double Consumtion { get; set; }
         public string PicturePath { get; set; }
         public string Comment { get; set; }
         [Ignore]
         public bool SNinDB { get; set; }
 
-        public Meter(int serialNumber, double consumtion, string picturePath, string comment)
+        public Meter(long serialNumber, double consumtion, string picturePath, string comment)
         {
             SerialNumber = serialNumber;
             Consumtion = consumtion;
@@ -25,7 +25,7 @@ namespace VVS.Model
             Comment = comment;
         }
 
-        public Meter(int serialNumber)
+        public Meter(long serialNumber)
         {
             SerialNumber = serialNumber;
         }
